🚚 moved ClipboardService into a single directory
