+++ libs/ui/shared/elements/src/lib/models/index.ts
@@ -1,5 +1,9 @@
+export type { } from '../services/clipboard';
+
 export type { IconModel } from './icon/icon.model';
+
 export type { ImageModel } from './image/image.model';
+
 export type { FlagModel } from './flag/flag.model';
 
 export type { LoaderConfigModel } from '../components/loader';
