+++ libs/ui/shared/elements/src/lib/services/clipboard/index.ts
@@ -1,0 +1,5 @@
+export type { } from './model/clipboard.model';
+
+export { } from './constant/clipboard.constants';
+
+export { ClipboardService } from './service/clipboard.service';
