+++ libs/ui/shared/elements/src/lib/services/clipboard/service/clipboard.service.ts
